videoPlayer: index.js -> video-player.js
